Add files via upload
The project mainly focusses on acquiring nonlinear data into linear basis function to fit a linear regression model to the data. We have used Gaussian function as a basis function. The main focus of the project is on the implementation of stochastic gradient descent from scratch over a nonlinear data.